Move to the parent folder which was empty
